refactor : separate enum class
